requirements.txt
pandas==2.2.3
numpy==1.26.4
matplotlib==3.9.2
fastapi==0.115.5
uvicorn==0.32.0
pydantic==2.9.2
#huggingface_hub==0.26.2
#pymongo==4.10.1
#torch==2.5.1
#transformers==4.46.2
requests==2.32.3
langchain==0.3.9
langchain_community==0.3.9
google-generativeai==0.8.3
langchain-google-genai==2.0.4
tiktoken==0.8.0
plotly==5.24.1
streamlit==1.40.2
duckduckgo-search==5.3.1b1
google-search-results==2.4.2
langchain_google_community==2.0.3
statsbombpy==1.14.0
mplsoccer==1.4.0
SPARQLWrapper==2.0.0
streamlit_extras==0.5.0
pyyaml==6.0.2
#langgraph